Consolidate callable, namespace and C-ABI semantic authorities
Integrate canonical code/frame/cell metadata, generated backend consumers, source-import and package custody, typed C-ABI ownership and numeric operand presence with regression sources and aligned public documentation. Source consolidation only: clean-source runtime/backend checks, retained codec replay and same-revision native/WASM acceptance remain pending; no release or donor-prune claim.

+++ runtime/molt-runtime/src/object/ops_slice.rs
@@ -397,7 +397,7 @@ pub extern "C" fn molt_dataclass_get(obj_bits: u64, index_bits: u64) -> u64 {
                         "field"
                     };
                     dec_ref_bits(_py, val);
-                    return attr_error(_py, "dataclass", name) as u64;
+                    return attr_error(_py, "dataclass", name);
                 }
                 return val;
             }
